% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_data_to_list.R
\name{base_url_download}
\alias{base_url_download}
\title{Helper function: base url for download}
\usage{
base_url_download()
}
\value{
base url for download of csv/zip files prepared by R package
}
\description{
Helper function: base url for download
}
