% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{columns_to_labels}
\alias{columns_to_labels}
\title{Create Text Labels from Data Frame Columns}
\usage{
columns_to_labels(data, columns, fmt = "\%s: \%s", sep = ", ")
}
\arguments{
\item{data}{data frame}

\item{columns}{names of columns from which to create labels}

\item{fmt}{format string passed to \code{\link{sprintf}}}

\item{sep}{separator (default: ", ")}
}
\value{
vector of character with as many elements as there are rows in data
}
\description{
Create Text Labels from Data Frame Columns
}
\examples{
data <- data.frame(number = 1:2, name = c("adam", "eva"), value = 3:4)
columns <- c("name", "value")
columns_to_labels(data, columns)
columns_to_labels(data, columns, fmt = "<p>\%s: \%s</p>", sep = "")
}
