% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_tables.R
\name{get_api_tables}
\alias{get_api_tables}
\title{Provide Tables of Wasserportal API Documentation}
\usage{
get_api_tables(name = NULL)
}
\arguments{
\item{name}{of element from the list of data frames to be selected. If this
argument is left blank (name = NULL), the default, the list of data frames
is returned.}
}
\value{
list of data frames or data frame specified by the \code{name} argument
}
\description{
The tables that appear in the API documentation of the wasserportal
(https://wasserportal.berlin.de/download/wasserportal_berlin_getting_data.pdf)
have been added to the wasserportal package. This function returns a list
of data frames with each element representing on of these tables.
}
\examples{
get_api_tables()

}
