% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater_data.R
\name{get_groundwater_data}
\alias{get_groundwater_data}
\title{Get Groundwater Data}
\usage{
get_groundwater_data(
  stations,
  groundwater_options = get_groundwater_options(),
  debug = TRUE,
  stations_list = NULL
)
}
\arguments{
\item{stations}{list as retrieved by \code{\link{get_stations}}.
Deprecated. Please use \code{stations_list} instead}

\item{groundwater_options}{as retrieved by \code{\link{get_groundwater_options}}}

\item{debug}{print debug messages (default: TRUE)}

\item{stations_list}{list of station metadata as returned by
\code{\link{get_stations}(type = "list")}}
}
\value{
list with elements "groundwater.level" and "groundwater.quality" data
frames
}
\description{
wrapper function to scrape all available raw data, i.e. groundwater
level and quality data and save in list
}
\examples{
\dontrun{
stations <- wasserportal::get_stations()
gw_data_list <- get_groundwater_data(stations)
str(gw_data_list)
}
}
