% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater_data.R
\name{get_groundwater_options}
\alias{get_groundwater_options}
\title{Helper function: get groundwater options}
\usage{
get_groundwater_options()
}
\value{
return available groundwater data options and prepare for being used
as input for \code{\link{get_groundwater_data}}
}
\description{
Helper function: get groundwater options
}
\examples{
get_groundwater_options()

}
