% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_variables.R
\name{get_station_variables}
\alias{get_station_variables}
\title{Helper function: get available station variables}
\usage{
get_station_variables(station_df)
}
\arguments{
\item{station_df}{data frame with one row per station and columns
"Messstellennummer", "Messstellenname" and additional columns each of which
represents a variable that is measured at that station. If the variable
columns contain the value "x" it means that the corresponding variable is
measured and the name of the column is contained in the returned vector of
variable names.}
}
\value{
returns names of available variables for station
}
\description{
Helper function: get available station variables
}
