% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations}
\usage{
get_stations(run_parallel = TRUE, n_cores = parallel::detectCores() - 1L)
}
\arguments{
\item{run_parallel}{default: TRUE}

\item{n_cores}{number of cores to use if \code{run_parallel = TRUE}.
Default: one less than the detected number of cores.}
}
\value{
list with general station "overview" (either as list "overview_list"
or as data.frame "overview_df") and a crosstable with information which
parameters is available per station ("x" if available, NA if not)
}
\description{
Get Stations
}
\examples{
stations <- wasserportal::get_stations(n_cores = 2L)
str(stations)

}
