% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surfacewater_qualities.R
\name{get_surfacewater_qualities}
\alias{get_surfacewater_qualities}
\title{Get Surface Water Quality for Multiple Monitoring Stations}
\usage{
get_surfacewater_qualities(station_ids, dbg = TRUE)
}
\arguments{
\item{station_ids}{vector with ids of multiple (or one) monitoring stations}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
data frame with water quality data for multiple monitoring stations
}
\description{
Get Surface Water Quality for Multiple Monitoring Stations
}
\examples{
\dontrun{
stations <- wasserportal::get_stations()
station_ids <- stations$overview_list$surface_water.quality$Messstellennummer
swq <- wasserportal::get_surfacewater_qualities(station_ids)
str(swq)
}
}
