% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surfacewater_quality.R
\name{get_surfacewater_quality}
\alias{get_surfacewater_quality}
\title{Get Surface Water Quality for One Monitoring Station}
\usage{
get_surfacewater_quality(station_id)
}
\arguments{
\item{station_id}{id of surface water measurement station}
}
\value{
data frame with water quality data for one monitoring station
}
\description{
Get Surface Water Quality for One Monitoring Station
}
\examples{
\dontrun{
stations <- wasserportal::get_stations()
station_id <- stations$overview_list$surface_water.quality$Messstellennummer[1]
swq <- wasserportal::get_surfacewater_quality(station_id)
str(swq)
}

}
