% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_surfacewater_data.R
\name{get_surfacewater_variables}
\alias{get_surfacewater_variables}
\title{Helper function: get surface water variables}
\usage{
get_surfacewater_variables()
}
\value{
vector with surface water variables
}
\description{
Helper function: get surface water variables
}
