% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_master_data.R
\name{get_wasserportal_master_data}
\alias{get_wasserportal_master_data}
\title{Wasserportal Berlin: get master data for a single station}
\usage{
get_wasserportal_master_data(master_url)
}
\arguments{
\item{master_url}{url with master data for single station as retrieved by
\code{\link{get_wasserportal_stations_table}}}
}
\value{
data frame with metadata for selected station
}
\description{
Wasserportal Berlin: get master data for a single station
}
\examples{

stations <- wasserportal::get_stations()

## GW Station
master_url <- stations$overview_list$groundwater.level$stammdaten_link[1]
get_wasserportal_master_data(master_url)

## SW Station
master_url <- stations$overview_list$surface_water.water_level$stammdaten_link[1]
get_wasserportal_master_data(master_url)

}
