% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_stations.R
\name{get_wasserportal_stations}
\alias{get_wasserportal_stations}
\title{Get Names and IDs of the Stations of wasserportal.berlin.de}
\usage{
get_wasserportal_stations(type = "quality")
}
\arguments{
\item{type}{one of "quality", "level", "flow"}
}
\description{
Get Names and IDs of the Stations of wasserportal.berlin.de
}
