% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_stations_table.R
\name{get_wasserportal_stations_table}
\alias{get_wasserportal_stations_table}
\title{Wasserportal Berlin: get stations overview table}
\usage{
get_wasserportal_stations_table(
  type = get_overview_options()$groundwater$level,
  url_wasserportal = wasserportal_base_url()
)
}
\arguments{
\item{type}{type of stations table to retrieve. Valid options defined in
\code{\link{get_overview_options}}, default: get_overview_options()$groundwater$level}

\item{url_wasserportal}{base url to Wasserportal berlin (default:
\code{\link{wasserportal_base_url}}}
}
\value{
data frame with master data of selected monitoring stations
}
\description{
Wasserportal Berlin: get stations overview table
}
\examples{
types <- wasserportal::get_overview_options()
str(types)
sw_l <- wasserportal::get_wasserportal_stations_table(type = types$surface_water$water_level)
str(sw_l)
}
