% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_variables.R
\name{get_wasserportal_variables}
\alias{get_wasserportal_variables}
\title{Get Names and IDs of the Variables of wasserportal.berlin.de}
\usage{
get_wasserportal_variables(station = NULL)
}
\arguments{
\item{station}{station id. If given, only variables that are available for
the given station are returned.}
}
\description{
Get Names and IDs of the Variables of wasserportal.berlin.de
}
