% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_to_csv.R
\name{list_data_to_csv}
\alias{list_data_to_csv}
\title{Helper function: list data to csv}
\usage{
list_data_to_csv(data_list, file_prefix = "")
}
\arguments{
\item{data_list}{data in list form}

\item{file_prefix}{file prefix (default: "")}
}
\value{
loops through list of data frames and uses list names as filenames
}
\description{
Helper function: list data to csv
}
