% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_to_csv_or_zip.R
\name{list_data_to_csv_or_zip}
\alias{list_data_to_csv_or_zip}
\title{Helper function: list data to csv}
\usage{
list_data_to_csv_or_zip(data_list, file_prefix = "", to_zip = FALSE)
}
\arguments{
\item{data_list}{data in list form}

\item{file_prefix}{file prefix (default: "")}

\item{to_zip}{convert to zip file (default: FALSE)}
}
\value{
loops through list of data frames and uses list names as filenames
}
\description{
Helper function: list data to csv
}
