% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_masters_data_to_csv.R
\name{list_masters_data_to_csv}
\alias{list_masters_data_to_csv}
\title{Helper function: list masters data to csv}
\usage{
list_masters_data_to_csv(masters_data_list)
}
\arguments{
\item{masters_data_list}{masters data in list form as retrieved by
\code{\link{get_stations}}\code{(type = "list")}}
}
\value{
loops through list of data frames and uses list names as filenames
}
\description{
Helper function: list masters data to csv
}
\examples{
\dontrun{
stations_list <- get_stations(type = "list")
masters_data_csv_files <- list_masters_data_to_csv(stations_list)
masters_data_csv_files
}
}
