% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_timeseries_data_to_zip.R
\name{list_timeseries_data_to_zip}
\alias{list_timeseries_data_to_zip}
\title{Helper function: list timeseries data to zip}
\usage{
list_timeseries_data_to_zip(timeseries_data_list)
}
\arguments{
\item{timeseries_data_list}{time series data in list form as retrieved by
\code{\link{get_groundwater_data}}}
}
\value{
loops through list of data frames and uses list names as filenames
}
\description{
Helper function: list timeseries data to zip
}
\examples{
\dontrun{
stations <- wasserportal::get_stations()

# Groundwater Time Series
gw_tsdata_list <- wasserportal::get_groundwater_data(stations)
gw_tsdata_files <- wasserportal::list_timeseries_data_to_zip(gw_tsdata_list)

# Surface Water Time Series
sw_tsdata_list <- wasserportal::get_daily_surfacewater_data(stations)
sw_tsdata_files <- wasserportal::list_timeseries_data_to_zip(sw_tsdata_list)
}
}
