% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read}
\alias{read}
\title{Helper function to read CSV}
\usage{
read(text, ...)
}
\arguments{
\item{text}{text}

\item{...}{\dots additional arguments passed to \code{\link[utils]{read.table}}}
}
\value{
data frame with values
}
\description{
Helper function to read CSV
}
