% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasserportal_raw.R
\name{read_wasserportal_raw}
\alias{read_wasserportal_raw}
\title{Read Wasserportal Raw}
\usage{
read_wasserportal_raw(
  variable,
  station,
  from_date,
  type = "single",
  include_raw_time = FALSE,
  handle = NULL,
  stations_crosstable
)
}
\arguments{
\item{variable}{variable}

\item{station}{station id}

\item{from_date}{start date}

\item{type}{one of "single", "daily", "monthly" (default: "single")}

\item{include_raw_time}{TRUE or FALSE (default: FALSE)}

\item{handle}{handle (default: NULL)}

\item{stations_crosstable}{sublist \code{crosstable} as retrieved from \link{\code{get_stations}}
i.e. \code{get_stations()$crosstable}}
}
\value{
????
}
\description{
Read Wasserportal Raw
}
