% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readPackageFile}
\alias{readPackageFile}
\title{Read CSV File from Package's "extdata" Folder}
\usage{
readPackageFile(file, ...)
}
\arguments{
\item{file}{file name (without path)}

\item{\dots}{additional arguments passed to \code{\link[utils]{read.csv}}}
}
\value{
data frame representing the content of \code{\link{file}}
}
\description{
Read CSV File from Package's "extdata" Folder
}
