% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_masters_data_to_list.R
\name{wp_masters_data_to_list}
\alias{wp_masters_data_to_list}
\title{Wasserportal Master Data: download and Import in R List}
\usage{
wp_masters_data_to_list(
  overview_list_names,
  target_dir = tempdir(),
  file_prefix = "stations_",
  is_zipped = FALSE
)
}
\arguments{
\item{overview_list_names}{names of elements in the list returned by
\code{\link{get_stations}(type = "list")}}

\item{target_dir}{target directory for downloading data (default:
tempdir())}

\item{file_prefix}{prefix given to file names}

\item{is_zipped}{are the data to be downloaded zipped (default:
FALSE)}
}
\value{
downloads csv master data from Wasserportal
}
\description{
Wasserportal Master Data: download and Import in R List
}
\examples{
\dontrun{
overview_list_names <- names(wasserportal::get_stations(type = "list"))
wp_masters_data_list <- wp_masters_data_to_list(overview_list_names)
}
}
