% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview_options.R
\name{get_overview_options}
\alias{get_overview_options}
\title{Wasserportal Berlin: get overview options for stations}
\usage{
get_overview_options()
}
\value{
list with shortcuts to station overview tables
(\verb{wasserportal.berlin.de/messwerte.php?anzeige=tabelle&thema=<shortcut>})
}
\description{
Wasserportal Berlin: get overview options for stations
}
\examples{
get_overview_options()

}
