% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_variables.R
\name{get_station_variables}
\alias{get_station_variables}
\title{Helper function: get available station variables}
\usage{
get_station_variables(station_df)
}
\arguments{
\item{station_df}{station_df}
}
\value{
returns names of available variables for station
}
\description{
Helper function: get available station variables
}
