% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations}
\usage{
get_stations(
  type = c("list", "data.frame", "crosstable"),
  run_parallel = TRUE,
  n_cores = parallel::detectCores() - 1L,
  debug = TRUE
)
}
\arguments{
\item{type}{vector of character describing the type(s) of output(s) to be
returned. Expected values (and default): \code{c("list", "data.frame",
  "crosstable")}. If only one value is given the data is returned in the
expected type. If more than one values are given, a list is returned with
one list element per type.}

\item{run_parallel}{default: TRUE}

\item{n_cores}{number of cores to use if \code{run_parallel = TRUE}.
Default: one less than the detected number of cores.}

\item{debug}{logical indicating whether or not to show debug messages}
}
\value{
list with general station "overview" (either as list "overview_list"
or as data.frame "overview_df") and a crosstable with information which
parameters is available per station ("x" if available, NA if not)
}
\description{
Get Stations
}
\examples{
stations <- wasserportal::get_stations(n_cores = 2L)
str(stations)

}
