% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations}
\usage{
get_stations(run_parallel = TRUE)
}
\arguments{
\item{run_parallel}{default: TRUE}
}
\value{
list with general station "overview" (either as list "overview_list"
or as data.frame "overview_df") and a crosstable with information which
parameters is available per station ("x" if available, NA if not)
}
\description{
Get Stations
}
\examples{
stations <- wasserportal::get_stations()
str(stations)

}
