% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_master_data.R
\name{get_wasserportal_master_data}
\alias{get_wasserportal_master_data}
\title{Wasserportal Berlin: get master data for a single station}
\usage{
get_wasserportal_master_data(
  station_id,
  url_wasserportal = wasserportal_base_url()
)
}
\arguments{
\item{station_id}{station_id}

\item{url_wasserportal}{base url to Wasserportal berlin (default:
wasserportal_base_url())}
}
\value{
data frame with metadata for
}
\description{
Wasserportal Berlin: get master data for a single station
}
\examples{
## GW Station
\dontrun{
get_wasserportal_master_data(station_id = 149)
}
## SW Station
get_wasserportal_master_data(station_id = 5865900)
}
