% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_masters_data.R
\name{get_wasserportal_master_data}
\alias{get_wasserportal_master_data}
\title{Wasserportal Berlin: get master data for a single station}
\usage{
get_wasserportal_master_data(master_url)
}
\arguments{
\item{master_url}{url with master data for single station as retrieved by
\code{\link{get_wasserportal_stations_table}}}
}
\value{
data frame with metadata for selected station
}
\description{
Wasserportal Berlin: get master data for a single station
}
\examples{
\dontrun{
stations_list <- wasserportal::get_stations(type = "list")

# GW Station
master_url <- stations_list \%>\%
  kwb.utils::selectElements("groundwater.level") \%>\%
  kwb.utils::selectColumns("stammdaten_link")[1L]

get_wasserportal_master_data(master_url)

# SW Station

# Reduce  to monitoring stations maintained by Berlin
master_urls <- stations_list \%>\%
  kwb.utils::selectElements("surface_water.water_level") \%>\%
  dplyr::filter(.data$Betreiber == "Land Berlin") \%>\%
  dplyr::pull(.data$stammdaten_link)

get_wasserportal_master_data(master_urls[1L])
}

}
