% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_masters_data.R
\name{get_wasserportal_masters_data}
\alias{get_wasserportal_masters_data}
\title{Wasserportal Berlin: get master data for a multiple stations}
\usage{
get_wasserportal_masters_data(station_ids, run_parallel = TRUE)
}
\arguments{
\item{station_ids}{station_ids}

\item{run_parallel}{default: TRUE}
}
\value{
data frame with metadata for selected station ids
}
\description{
Wasserportal Berlin: get master data for a multiple stations
}
\examples{
station_ids <- 1:4
parallel::detectCores()
system.time(get_wasserportal_masters_data(station_ids))
system.time(get_wasserportal_masters_data(station_ids,
                                          run_parallel = FALSE))

}
