% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasserportal_raw.R
\name{read_wasserportal_raw}
\alias{read_wasserportal_raw}
\title{Read Wasserportal Raw}
\usage{
read_wasserportal_raw(
  variable,
  station,
  from_date,
  type = "single",
  include_raw_time = FALSE,
  handle = NULL,
  stations_crosstable,
  api_version = 2L
)
}
\arguments{
\item{variable}{variable}

\item{station}{station id}

\item{from_date}{start date}

\item{type}{one of "single", "daily", "monthly" (default: "single")}

\item{include_raw_time}{TRUE or FALSE (default: FALSE)}

\item{handle}{handle (default: NULL)}

\item{stations_crosstable}{data frame as returned by
\code{\link{get_stations}(type = "crosstable")}}

\item{api_version}{1 integer number representing the version of
wasserportal's API. 1L: before 2023, 2L: since 2023. Default: 2L}
}
\value{
????
}
\description{
Read Wasserportal Raw
}
