% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasserportal_raw_gw.R
\name{read_wasserportal_raw_gw}
\alias{read_wasserportal_raw_gw}
\title{read_wasserportal_raw_gw}
\usage{
read_wasserportal_raw_gw(
  station = 149,
  stype = "gws",
  type = "single_all",
  from_date = "",
  include_raw_time = FALSE,
  handle = NULL
)
}
\arguments{
\item{station}{station id}

\item{stype}{"gws" or "gwq"}

\item{type}{"single" or "single_all" (if stype = "gwq")}

\item{from_date}{(default: "")}

\item{include_raw_time}{default: FALSE}

\item{handle}{default: NULL}
}
\value{
data.frame with values
}
\description{
read_wasserportal_raw_gw
}
\examples{
\dontrun{
read_wasserportal_raw_gw(station = 149, stype = "gws")
read_wasserportal_raw_gw(station = 149, stype = "gwq")
}
}
