% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_master_data.R
\name{wasserportal_base_url}
\alias{wasserportal_base_url}
\title{Helper function: Base Url of Berlin Wassersportal}
\usage{
wasserportal_base_url()
}
\value{
string with base url of Berlin Wasserportal
}
\description{
Helper function: Base Url of Berlin Wassersportal
}
