% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_time-series_data_to_list.R
\name{wp_timeseries_data_to_list}
\alias{wp_timeseries_data_to_list}
\title{Wasserportal Time Series Data: download and Import in R List}
\usage{
wp_timeseries_data_to_list(
  overview_list_names,
  target_dir = tempdir(),
  is_zipped = TRUE
)
}
\arguments{
\item{target_dir}{target directory for downloading data (default:
tempdir())}

\item{is_zipped}{are the data to be downloaded zipped (default:
TRUE)}

\item{overview_list}{names of "overview_list" as retrieved by
\code{\link{get_stations}}}
}
\value{
downloads (zipped) data from wasserportal
}
\description{
Wasserportal Time Series Data: download and Import in R List
}
\examples{
\dontrun{
stations <- wasserportal::get_stations()
overview_list_names <- names(stations$overview_list)
wp_timeseries_data_list <- wp_timeseries_data_to_list(overview_list_names)
}
}
